refactor: move L1BlockId from strata-state to strata-primitives
